fix:fix

+++ hw4/task10.ts
@@ -5,7 +5,7 @@ const readline10 = require("readline-sync");
 //Дана произвольное число с плавающей точкой. Реализовать методы округления данного числа до ближайшего меньшего, 
 //большего аргумента и ближайшего целого числа.
 
-const num10 = parseInt(readline10.question("Type a number "));
+const num10 = parseFloat(readline10.question("Type a number "));
 const floor = Math.floor(num10);
 const ceil = Math.ceil(num10);
 const round = Math.round(num10);
