+++ hw4/task9.ts
@@ -5,7 +5,7 @@ const readline9 = require("readline-sync");
 //Дана произвольное число. Реализовать метод проверяющий является ли это число целочисленным или нет. 
 //Если является - вывести его, если нет - округлить до целой части.
 
-const someNum = parseInt(readline9.question("Type a number "));
+const someNum = parseFloat(readline9.question("Type a number "));
 
 if(Number.isInteger(someNum))
 console.log(someNum);
